+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Minimal Go version is now 1.23
 
+### Fixed
+
+- One HTTP PUT with Content-Length in cmaf-ingest
+- Fixed typo in urlgen template
+- Lost error reporting in initial segment parsing
+
 ## [1.7.0] - 2025-01-17
 
 ### Changed