+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - HEVC encryption support
+- AC-4 audio support including encryption
+- AC-3 and EC-3 encryption support
 
 ### Changed